Rename api/auth/util.py to api/auth/auth_util.py
